% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_mappings.R
\name{add_slope}
\alias{add_slope}
\title{Add length averaged slope to flowpaths}
\usage{
add_slope(flowpaths)
}
\arguments{
\item{flowpaths}{sf LINESTRING}
}
\value{
sf LINSETRING
}
\description{
Add length weighted slopes to refactored/aggregated flowpaths
using the base NHDPlus slope estimates
}
