% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_lstm.R
\name{aggregate_basin_attributes}
\alias{aggregate_basin_attributes}
\title{Aggregated LSTM variables}
\usage{
aggregate_basin_attributes(
  gpkg,
  catchment_name,
  geo_dir,
  years = 3,
  precision = 9,
  out_file = NULL
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{catchment_name}{the layer name of the aggregation units}

\item{geo_dir}{the geogroids geo directory}

\item{years}{number of years of PPT and Temperture to process}

\item{precision}{the precision of the computations}

\item{out_file}{the CSV file path to write the results to}
}
\value{
the out_file path
}
\description{
Aggregated LSTM variables
}
