% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cfe_noahowp.R
\name{aggregate_cfe}
\alias{aggregate_cfe}
\title{Aggregated CFE variables frome NWM}
\usage{
aggregate_cfe(
  gpkg,
  nwm_dir,
  catchment_name,
  flowline_name,
  single_layer = FALSE,
  precision = 9,
  out_file = NULL
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{nwm_dir}{the NWM directory}

\item{catchment_name}{the layer name of the aggregation units}

\item{flowline_name}{the layer name of the flowpath units}

\item{single_layer}{should only the top layer of a multilayer parameter be processed?}

\item{precision}{the precision of the computations}

\item{out_file}{the location to write output}
}
\description{
Aggregated CFE variables frome NWM
}
