% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_nlcd.R
\name{aggregate_nlcd}
\alias{aggregate_nlcd}
\title{Aggregated NLCD variables}
\usage{
aggregate_nlcd(
  gpkg,
  catchment_name,
  imperv_path = TRUE,
  lc_path = TRUE,
  precision = 9,
  out_file = NULL
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{catchment_name}{the layer name of the aggregation units}

\item{imperv_path}{path to NLCD impervious IMG file}

\item{lc_path}{path to NLCD landcover IMG file}

\item{precision}{the precision of the computations}

\item{out_file}{the file to write to (CSV using data.table fwrite)}
}
\description{
Aggregated NLCD variables
}
