% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_nwm.R
\name{aggregate_nwm_params}
\alias{aggregate_nwm_params}
\title{Aggregated NWM variables}
\usage{
aggregate_nwm_params(
  gpkg,
  nwm_dir,
  catchment_name,
  flowline_name,
  precision = 9,
  out_dir = NULL
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{nwm_dir}{the geogroids geo directory}

\item{catchment_name}{the layer name of the aggregation units}

\item{flowline_name}{the layer name of the flowpath units}

\item{precision}{the precision of the computations}

\item{out_dir}{the directory to write files to}
}
\description{
Aggregated NWM variables
}
