% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelists.R
\name{build_edge_lists}
\alias{build_edge_lists}
\title{Build catchment, flowpath and waterbody edgelists
Returns catchment, flowpath and waterbody edgelists. If \code{outdir} is supplied,
then JSON files are written there.}
\usage{
build_edge_lists(
  flowpaths,
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-",
  wb_prefix = "wb-",
  terminal_wb_prefix = "twb-",
  catchment_prefix = "cat-",
  cutoff = 100000000,
  outdir
)
}
\arguments{
\item{flowpaths}{sf data.frame containing hyRefactor or hyAggregate output.}

\item{nexus_prefix}{character prefix for nexus IDs}

\item{terminal_nexus_prefix}{character prefix for terminal nexus IDs}

\item{wb_prefix}{character prefix for waterbody IDs}

\item{terminal_wb_prefix}{character prefix for terminal waterbody IDs}

\item{catchment_prefix}{character prefix for catchment IDs}

\item{cutoff}{terminal IDs begin above a defiend threshold}

\item{outdir}{optional. Directory to write JSON files}
}
\value{
list of edgelist
}
\description{
Build catchment, flowpath and waterbody edgelists
Returns catchment, flowpath and waterbody edgelists. If \code{outdir} is supplied,
then JSON files are written there.
}
