% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_mappings.R
\name{build_lake_params}
\alias{build_lake_params}
\title{Build Lake Parms for NGen}
\usage{
build_lake_params(
  gpkg,
  catchment_name,
  flowline_name,
  waterbody_gpkg = "/Users/mjohnson/Downloads/nhdplus_waterbodies.gpkg",
  nwm_dir
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{catchment_name}{the layer name of the aggregation units}

\item{flowline_name}{the layer name of the flowpath units}

\item{waterbody_gpkg}{geopackage containing CONUS NHD waterboides}

\item{nwm_dir}{a directory to NWM files}
}
\description{
Map the NWM LAKEPARM_CONUS.nc file to an aggregated NGen Network
}
