% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_release_dir.R
\name{build_release_directory}
\alias{build_release_directory}
\title{Build Release Directory Structure}
\usage{
build_release_directory(root = "")
}
\arguments{
\item{root}{the root path}
}
\value{
list of directory paths
}
\description{
in a root directory, create a graph, spatial, parameters,
and crosswalks sub directories
}
