% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_mappings.R
\name{catchment_waterbody_interaction}
\alias{catchment_waterbody_interaction}
\title{Quantify area of waterbody/land per catchment
Ngen needs a land area percentages to drive rainfall-runoff processes. Provided a Ngen geopackage,
and NHD waterbody geopackage, these are computed.}
\usage{
catchment_waterbody_interaction(
  gpkg = "/Users/mjohnson/Downloads/2021-10-28/spatial/hydrofabric.gpkg",
  catchment_name = "catchments",
  waterbody_gpkg = "/Users/mjohnson/Downloads/nhdplus_waterbodies.gpkg",
  nwm_dir = "/Users/mjohnson/Downloads/"
)
}
\arguments{
\item{gpkg}{Geopackage containing catchment data.}

\item{waterbody_gpkg}{Geopackage of NHD waterbodies}

\item{nwm_dir}{if supplied, only waterbodies included in the NWM Routelink file will be
included in the area calculations. From the NWM directory, the Routelink file is searched for.}
}
\value{

}
\description{
Quantify area of waterbody/land per catchment
Ngen needs a land area percentages to drive rainfall-runoff processes. Provided a Ngen geopackage,
and NHD waterbody geopackage, these are computed.
}
