% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_mappings.R
\name{check_file}
\alias{check_file}
\title{File Check}
\usage{
check_file(filename, overwrite = FALSE)
}
\arguments{
\item{filename}{character. Output filename.}

\item{overwrite}{logical. If TRUE, filename is overwritten}
}
\value{
logical
}
\description{
Check if a file exists and/or should be over ridden.
}
