% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{file.path.build}
\alias{file.path.build}
\title{Build a file.path explicitly}
\usage{
file.path.build(...)
}
\arguments{
\item{flowpaths}{sf object (LINESTRING)}

\item{catchments}{sf object (POLYGON)}

\item{nexus}{sf object (POINT)}

\item{spatial_path}{where to right the spatial path}
}
\value{

}
\description{
Build a file.path explicitly
}
