% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nexus.R
\name{get_nex_locations}
\alias{get_nex_locations}
\title{Get Nexus Locations}
\usage{
get_nex_locations(fp, term_cut)
}
\arguments{
\item{fp}{a sf flowpath object}
}
\value{

}
\description{
Get Nexus Locations
}
