% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{gpkg_layers}
\alias{gpkg_layers}
\title{Write NGEN spatial products}
\usage{
gpkg_layers(path, goal, pattern = NULL)
}
\arguments{
\item{flowpaths}{sf object}

\item{catchments}{sf object}

\item{nexus}{sf object}

\item{spatial_path}{where to right the spatial path}
}
\value{

}
\description{
Write NGEN spatial products
}
