% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{length_average_routelink}
\alias{length_average_routelink}
\title{Length Average Routelink Variables}
\usage{
length_average_routelink(
  flowpaths,
  rl_vars = c("link", "Qi", "MusK", "MusX", "n", "So", "ChSlp", "BtmWdth", "time",
    "Kchan", "nCC", "TopWdthCC", "TopWdth"),
  rl_path
)
}
\arguments{
\item{flowpaths}{sf LINESTRING}

\item{rl_vars}{routelink variables}

\item{rl_path}{routelink path}
}
\value{
sf LINESTRING object
}
\description{
Length Average Routelink Variables
}
