% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_average_routelink.R,
%   R/update_functions.R
\name{length_average_routlink}
\alias{length_average_routlink}
\title{Assign Length Weighted Flowpath Attributes from Routelink file}
\usage{
length_average_routlink(
  flowpaths,
  rl_vars = c("link", "Qi", "MusK", "MusX", "n", "So", "ChSlp", "BtmWdth", "time",
    "Kchan", "nCC", "TopWdthCC", "TopWdth"),
  rl_path
)

length_average_routlink(
  flowpaths,
  rl_vars = c("link", "Qi", "MusK", "MusX", "n", "So", "ChSlp", "BtmWdth", "time",
    "Kchan", "nCC", "TopWdthCC", "TopWdth"),
  rl_path
)
}
\arguments{
\item{flowpaths}{sf LINESTRING}

\item{rl_vars}{routelink variables}

\item{rl_path}{routelink path}
}
\value{
flowpaths with length weighted RouteLink Attributes

sf LINESTRING object
}
\description{
Assign Length Weighted Flowpath Attributes from Routelink file

Length Average Routelink Variable
}
