% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_nwm.R
\name{map_vars}
\alias{map_vars}
\title{Map Variables}
\usage{
map_vars(r, w, file = NULL, precision = 9, FUN = "mean", verbose = TRUE)
}
\arguments{
\item{r}{a file path, terra SpatRaster or raster raster object}

\item{w}{a weighting file computed with zonal::weighting_grid}

\item{file}{a CSV path to write data to, if NULL data.table object is returned}

\item{precision}{computational precision of the zonal calculations}

\item{FUN}{a summary function (see zonal::execute_zonal)}

\item{verbose}{should messages be emitted?}
}
\value{
file path or data.table
}
\description{
Wrapper to extract, rename and return or write a csv of zonal variables
}
