% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{refactor_wrapper}
\alias{refactor_wrapper}
\title{Refactoring Wrapper}
\usage{
refactor_wrapper(
  flowpaths,
  catchments,
  events = NULL,
  avoid = NULL,
  split_flines_meters = 10000,
  collapse_flines_meters = 1000,
  collapse_flines_main_meters = 1000,
  cores = 1,
  facfdr = NULL,
  routing = NULL,
  keep = 0.9,
  outfile
)
}
\arguments{
\item{flowpaths}{Reference flowline features}

\item{catchments}{Reference catchment features}

\item{events}{data.frame containing events}

\item{avoid}{integer vector of COMIDs to be excluded from collapse modifications.}

\item{split_flines_meters}{numeric the maximum length flowpath desired in the output.}

\item{collapse_flines_meters}{numeric the minimum length of inter-confluence flowpath desired in the output.}

\item{collapse_flines_main_meters}{numeric the minimum length of between-confluence flowpaths.}

\item{cores}{integer number of cores to use for parallel execution}

\item{facfdr}{path to directory with flow direction and flow accumulation \code{SpatRast}. If NULL (default) then catchments are NOT reconciled.}

\item{routing}{path to National Water Model RouteLink file. If NULL (default) then routing parameters are NOT added to the refactroed flowlines.}

\item{keep}{proportion of points to retain in geometry simplification (0-1; default 0.05). See ms_simplify. If NULL, then no simplification will be executed.}

\item{outfile}{path to geopackage to write refactored_flowlines, and if facfdr != NULL, refactored catchments.}
}
\value{
data to the specified gpkg
}
\description{
A wrapper around hyRefactor refactor_nhdplus and
reconcile_catchment_divides that optionally adds routing parameters to the
flowpath objects.
}
