% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_functions.R
\name{write_hydrofabric}
\alias{write_hydrofabric}
\title{Write Core Hydrofabric Files
The core hydrofabric files for NextGen include (A) a flowpath edgelist, (B) a catchment geojson,
(C) a nexus geojson, and a geopackage which includes the catchments POLYGONs,
flowpaths LINESTRINGs, and nexus POINTs}
\usage{
write_hydrofabric(
  gpkg,
  catchment_name,
  flowpath_name,
  ngen_pu,
  spatial_path,
  parameter_path,
  waterbody_prefix = "wb-",
  catchment_prefix = "cat-",
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-",
  overwrite = FALSE
)
}
\arguments{
\item{gpkg}{a filename of a gpkg to process}

\item{catchment_name}{the name of the catchment layer}

\item{flowpath_name}{the name of the flowpath layer}

\item{ngen_pu}{the name of the current processing unit}

\item{spatial_path}{path to spat}

\item{parameter_path}{path to parameter set}

\item{waterbody_prefix}{waterbody prefix}

\item{catchment_prefix}{catchment prefix}

\item{nexus_prefix}{nexus prefix}

\item{terminal_nexus_prefix}{terminal nexus prefix}

\item{overwrite}{logical. overwrite existing files?}
}
\value{
path to the resulting hydrofabric
}
\description{
Write Core Hydrofabric Files
The core hydrofabric files for NextGen include (A) a flowpath edgelist, (B) a catchment geojson,
(C) a nexus geojson, and a geopackage which includes the catchments POLYGONs,
flowpaths LINESTRINGs, and nexus POINTs
}
