% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json_files.R
\name{write_lakes_json}
\alias{write_lakes_json}
\title{Write lake params to JSON}
\usage{
write_lakes_json(df, outfile)
}
\arguments{
\item{df}{output of \code{build_lake_params}}

\item{outfile}{file path to write to}
}
\value{
file path
}
\description{
Write lake params to JSON
}
