% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{write_nextgen_spatial}
\alias{write_nextgen_spatial}
\title{Write NGEN spatial products}
\usage{
write_nextgen_spatial(flowpaths, catchments, nexus, spatial_path)
}
\arguments{
\item{flowpaths}{sf object}

\item{catchments}{sf object}

\item{nexus}{sf object}

\item{spatial_path}{where to right the spatial path}
}
\value{

}
\description{
Write NGEN spatial products
}
