% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json_files.R
\name{write_nwis_crosswalk}
\alias{write_nwis_crosswalk}
\title{Write NWIS mapping to JSON
Takes a geopackage or flowline layer, extracts the flowpath features,
and builds a lengthweighted average of all RouteLink variables using \code{length_average_routlink}.
The output data as written as JSON to the supplied outfile path.}
\usage{
write_nwis_crosswalk(
  gpkg,
  flowline_name,
  flowpaths = NULL,
  fl_name = "flowpaths",
  nwm_dir = "/Users/mjohnson/Downloads/",
  outfile
)
}
\arguments{
\item{gpkg}{geopackage with flowline features}

\item{fl_name}{name of flowline feature layer}

\item{nwm_dir}{directory to NWM base files. Searched for a file called \code{Routelink_CONUS.nc}}

\item{outfile}{file path to write to}
}
\value{
file path
}
\description{
Write NWIS mapping to JSON
Takes a geopackage or flowline layer, extracts the flowpath features,
and builds a lengthweighted average of all RouteLink variables using \code{length_average_routlink}.
The output data as written as JSON to the supplied outfile path.
}
