% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json_files.R
\name{write_waterbody_json}
\alias{write_waterbody_json}
\title{Write routing params to JSON
Takes a geopackage, extracts the flowpath features, and builds a lengthweighted average of
all RouteLink variables using \code{length_average_routlink}. The output data as written as JSON
to the supplied outfile path.}
\usage{
write_waterbody_json(fps, outfile)
}
\arguments{
\item{outfile}{file path to write to}

\item{gpkg}{geopackage with flowline features}

\item{fl_name}{name of flowline feature layer}

\item{nwm_dir}{directory to NWM base files. Searched for a file called \code{Routelink_CONUS.nc}}
}
\value{
file path
}
\description{
Write routing params to JSON
Takes a geopackage, extracts the flowpath features, and builds a lengthweighted average of
all RouteLink variables using \code{length_average_routlink}. The output data as written as JSON
to the supplied outfile path.
}
