% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir_mappings.R
\name{build_lake_params}
\alias{build_lake_params}
\title{Build Lake Params for NGen}
\usage{
build_lake_params(
  gpkg,
  flowline_name,
  flowpaths = NULL,
  waterbody_gpkg = "/Users/mjohnson/Downloads/nhdplus_waterbodies.gpkg",
  nwm_dir,
  outfile = NULL
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{flowline_name}{the layer name of the flowpath units}

\item{waterbody_gpkg}{geopackage containing CONUS NHD waterbodies}

\item{nwm_dir}{a directory to NWM files}

\item{outfile}{optional. Path to JSON file to write outputs using \code{write_lakes_json}}

\item{catchment_name}{the layer name of the aggregation units}
}
\description{
Map the NWM LAKEPARM_CONUS.nc file to an aggregated NGen Network
}
