% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyRelease-package.R
\docType{package}
\name{hyRelease-package}
\alias{hyRelease}
\alias{hyRelease-package}
\title{hyRelease: Build NGen Hydrofabric Releases}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikejohnson51/hyRelease}
  \item Report bugs at \url{https://github.com/mikejohnson51/hyRelease/issues}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{mikecp11@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
