% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_average_routelink.R
\name{length_average_routlink}
\alias{length_average_routlink}
\title{Assign Length Weighted Flowpath Attributes from Routelink file}
\usage{
length_average_routlink(
  flowpaths,
  rl_vars = c("link", "Qi", "MusK", "MusX", "n", "So", "ChSlp", "BtmWdth", "time",
    "Kchan", "nCC", "TopWdthCC", "TopWdth"),
  rl_path
)
}
\arguments{
\item{flowpaths}{flowpaths with a\code{member_COMID} attribute}

\item{rl_vars}{RouteLink variable to append}

\item{rl_path}{Path to RouteLink}
}
\value{
flowpaths with length weighted RouteLink Attributes
}
\description{
Assign Length Weighted Flowpath Attributes from Routelink file
}
