% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_functions.R
\name{write_waterbody_json}
\alias{write_waterbody_json}
\title{Write routing params to JSON
Takes a geopackage, extracts the flowpath features, and builds a length weighted average of
all RouteLink variables using \code{length_average_routlink}. The output data ss written as JSON
to the supplied outfile path.}
\usage{
write_waterbody_json(gpkg, flowpath_name, nwm_dir, outfile)
}
\arguments{
\item{gpkg}{geopackage with flowline features}

\item{flowpath_name}{name of flowline feature layer}

\item{nwm_dir}{directory to NWM base files. Searches for a file called \code{Routelink_CONUS.nc}}

\item{outfile}{file path to write to}
}
\value{
file path
}
\description{
Write routing params to JSON
Takes a geopackage, extracts the flowpath features, and builds a length weighted average of
all RouteLink variables using \code{length_average_routlink}. The output data ss written as JSON
to the supplied outfile path.
}
